added token verification to importActivities route

+++ backend/src/index.ts
@@ -7,12 +7,14 @@ import { ActivityEntry, NewUserActivity, UserActivity } from "../models/UserActi
 import ActivityRoute from "../routes/Activity";
 import { generateRandomColor, getTimeFromLongString } from "../../calendar/src/utils/helpers";
 import { ObjectId } from "bson";
+import { checkToken } from "../utils/helpers";
 
 // The Worker's environment bindings
 type Bindings = {
   ATLAS_APPID: string;
   ATLAS_APIKEY: string;
   PRIVATE_KEY: string; // private key used to sign jwt tokens
+  JWT_SECRET: string;
 };
 
 const app = new Hono<{ Bindings: Bindings }>();
@@ -103,11 +105,9 @@ app.post('/userColors', async (c) => {
 
 // Route to process the weekList and create new activities
 app.post('/importActivities', async (c) => {
-  const { userId, weekList }: { userId: ObjectId, weekList: any[] } = await c.req.json();
-
   App = App || new Realm.App(c.env.ATLAS_APPID);
   const credentials = Realm.Credentials.apiKey(c.env.ATLAS_APIKEY);
-  const user = await App.logIn(credentials); // Attempt to authenticate
+  const user = await App.logIn(credentials);
   const client = user.mongoClient("mongodb-atlas");
   const userCollection = client
     .db("calendar")
@@ -115,6 +115,28 @@ app.post('/importActivities', async (c) => {
   const activityCollection = client
     .db("calendar")
     .collection<UserActivity>("activity");
+
+  const cookieHeader = c.req.header("Cookie");
+  if (!cookieHeader) {
+    c.status(400);
+    return c.json({ message: "no cookies found" });
+  }
+
+  const cookies = cookieHeader.split(";").map((cookie) => cookie.trim());
+  let token = cookies.find((cookie) => cookie.startsWith(`token=`));
+  if (!token) {
+    c.status(400);
+    return c.json({ message: "no token found" });
+  }
+  token = token.split("=")[1].trim();
+
+  const id = await checkToken(token, c.env.JWT_SECRET);
+  if (!id) {
+    c.status(400);
+    return c.json({ message: "bad token" });
+  }
+  const { userId, weekList }: { userId: ObjectId, weekList: any[] } = await c.req.json();
+
   // 1. Erase all existing activities
   console.log(`there are: ${(await activityCollection.find({ userId: new ObjectId(userId.toString()) })).length} documents in activity collection`);
   await activityCollection.deleteMany({ userId: new ObjectId(userId.toString()) });
